move coo code out of ruby_sparse.c
